Added Scene
씬 로딩 화면 추가,
클리어 화면 연결

+++ Assets/Scripts/TrainingSystem/TrainingStructList.cs
@@ -17,7 +17,7 @@ public class TrainingStructList : MonoBehaviour
 
     [Header("Variables")]
     public int _percentage;
-    public int _trainDelayTime;
+    public float _trainDelayTime;
     public bool isUsingItem = false;
 
     int percentChkMinScore;
